+++ src/Fable.Transforms/Python/PythonCompiler.fs
@@ -132,7 +132,7 @@ type PythonCompiler(com: Compiler) =
         member _.GetImplementationFile(fileName) = com.GetImplementationFile(fileName)
 
         member _.GetRootModule(fileName) = com.GetRootModule(fileName)
-        member _.TryGetEntity(fullName) = com.TryGetEntity(fullName)
+        member _.TryGetEntity(entRef) = com.TryGetEntity(entRef)
         member _.GetInlineExpr(fullName) = com.GetInlineExpr(fullName)
 
         member _.AddWatchDependency(fileName) = com.AddWatchDependency(fileName)